% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_checkers.R
\name{check_medium_cv}
\alias{check_medium_cv}
\title{Menu functinon to check that sampled medium term is in controlled vocabulary}
\usage{
check_medium_cv(db, sampledmedium)
}
\arguments{
\item{db}{database connection}

\item{sampledmedium}{term to check}
}
\value{
sampled medium
}
\description{
Menu functinon to check that sampled medium term is in controlled vocabulary
}
\examples{
# medium <- check_medium_cv(db, "air)
}
