% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_checkers.R
\name{check_methodtype_cv}
\alias{check_methodtype_cv}
\title{Menu function to check that method type is in controlled vocabulary}
\usage{
check_methodtype_cv(db, methodtypecv)
}
\arguments{
\item{db}{database connection}

\item{methodtypecv}{method type to check}
}
\value{
method type
}
\description{
Menu function to check that method type is in controlled vocabulary
}
\examples{
# methodcode <- check_methodtype_cv()
}
