% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_samplingfeaturecodes}
\alias{check_samplingfeaturecodes}
\title{Compare sampling feature names to existing ones}
\usage{
check_samplingfeaturecodes(new_codes, db = db)
}
\arguments{
\item{new_codes}{vector of names to compare with samplingfeaturecode vector}

\item{db}{database connection (defaults to db)}
}
\value{
data frame with new_codes and in_db column
}
\description{
Returns a data frame specifying whether each element of new_codes exists in samplingfeatures
}
\examples{
\dontrun{
#
check_samplingfeaturecodes(new_codes = new_sites$sites)
}

}
