% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_variables_list}
\alias{check_variables_list}
\title{Check that the names of a variables list are in a database}
\usage{
check_variables_list(db, variables)
}
\arguments{
\item{db}{database connection}

\item{variables}{named list of variables}
}
\value{
messages about each new variable entered
}
\description{
Check that the names of a variables list are in a database
}
\examples{
#check_variables_list(db, varables = vars_list)
}
