% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sqlite.R
\name{connect_sqlite}
\alias{connect_sqlite}
\title{Connect to an existing ODM2 sqlite database}
\usage{
connect_sqlite(filename = "odm2", dir = ".")
}
\arguments{
\item{filename}{Filename for sqlite database file. Defaults to "odm2"}

\item{dir}{Name of existing directory where database file is stored. Defaults to current working directory}
}
\value{
A database connection object
}
\description{
Connect to an existing ODM2 sqlite database
}
\details{
This function returns a database connection object to be used in subsequent inserts and queries.
}
\examples{
# create_sqlite(filename = "odm2", dir = tempdir())
db <- connect_sqlite(filename = "odm2")
}
