% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_add_annotation.R
\name{db_add_annotation}
\alias{db_add_annotation}
\title{Add a new annotation}
\usage{
db_add_annotation(annotationtypecv, annotationtext)
}
\arguments{
\item{annotationtypecv}{annotation type from controlled vocab}

\item{annotationtext}{text string of annotation}
}
\value{
TRUE if successful
}
\description{
Add a new annotation
}
\examples{
#' \dontrun{
db_add_annotaion("samplingfeature annotation", "primray sites")
}
}
