% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_add_method.R
\name{db_add_method}
\alias{db_add_method}
\title{Add a new method}
\usage{
db_add_method(methodname, methodcode, methodtypecv, methoddescription)
}
\arguments{
\item{methodname}{full name of method}

\item{methodcode}{short codename for method}

\item{methodtypecv}{method type from controlled vocabulary}

\item{methoddescription}{longer description}
}
\value{
TRUE if successful
}
\description{
Add a new method
}
\examples{
\dontrun{
db_add_method("new method", "new", "Field Activity", "doing stuf")
}
}
