% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_annotate.R
\name{db_annotate}
\alias{db_annotate}
\title{Insert new annotation}
\usage{
db_annotate(db, object, annotationtext = NULL, type = "Site group",
  annotationcode = NULL)
}
\arguments{
\item{db}{database connecton object}

\item{object}{An existing object in the database to annotate such as a site or specimen code.}

\item{annotationtext}{Text of annotation}

\item{type}{annotation type from [controlled vocab](http://vocabulary.odm2.org/annotationtype/)}

\item{annotationcode}{annotation codename}
}
\value{
TRUE if successful
}
\description{
Insert new annotation
}
\details{
This function adds an existing annotation to an existing feature in the database
such as a site code, sampling feature, or action. Once objects are annotated, they can be
more easily queried as a group based on these labels or tags that are not otherwise
represented in the database structure.
Either an annotation code or annotation text must be provided to identify the annotation
}
\examples{
# db_annotate(db, object = "Site 001", annotationtext = "Riparian wells")
}
