% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_annotation.R
\name{db_describe_annotation}
\alias{db_describe_annotation}
\title{Add a new annotation text}
\usage{
db_describe_annotation(db, annotationtypecv, annotationtext = NULL,
  annotationcode = NULL)
}
\arguments{
\item{db}{database connection object}

\item{annotationtypecv}{annotation type from [controlled vocab](http://vocabulary.odm2.org/annotationtype/)}

\item{annotationtext}{text string of annotation}

\item{annotationcode}{Optional short codename for annotation}
}
\value{
TRUE if successful
}
\description{
Add a new annotation text
}
\details{
Use this function to add new annotations such as the name of a Site group.
}
\examples{
#db <- create_sqlite(dir = tempdir())
#db_describe_annotaion(db, "Site group", "Riparian wells")
#db_describe_annotaion(db, "Specimen group", "January sampling campaign", annotationcode = "Jan")

}
\seealso{
Other describe functions: \code{\link{db_describe_equipment}},
  \code{\link{db_describe_method}},
  \code{\link{db_describe_organization}},
  \code{\link{db_describe_person}},
  \code{\link{db_describe_site}},
  \code{\link{db_describe_variable}}
}
\concept{describe functions}
