% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_equipment.R
\name{db_describe_equipment}
\alias{db_describe_equipment}
\title{Add a piece of equipment to the equipment table}
\usage{
db_describe_equipment(db, equip_name, serial_no, model_name, vendor,
  owner_first, equipment_type = "Sensor", owner_last = NULL,
  owner_email = NULL, manufacturer = NULL,
  purchase_date = Sys.Date())
}
\arguments{
\item{db}{database connection}

\item{equip_name}{a unique short code name of the instrument}

\item{serial_no}{serial number of equipment}

\item{model_name}{model name of equipment}

\item{vendor}{organization that sold the product}

\item{owner_first}{first name of person who owns equipment}

\item{equipment_type}{type of equipment, defaults to sensor}

\item{owner_last}{last name of person who owns equipment,
required if the person is not in the people table already}

\item{owner_email}{email of person who owns equipment,
required if the person is not in the people table arelady}

\item{manufacturer}{company that makes the model of the equipment,
required if the model is not in the equipmentmodels table}

\item{purchase_date}{date that the equipment was purchased,
defaults to current date}
}
\value{
TRUE if successful
}
\description{
Add a piece of equipment to the equipment table
}
\examples{
# db <- create_sqlite()
# OR
# db <- connect_sqlite()
# db_describe_equipment(db, "unit1", "001", "model1",
# "vendor name", "Wendy", "Sensor", "Wetland", "ww email", "hobo")
}
\seealso{
Other describe functions: \code{\link{db_describe_annotation}},
  \code{\link{db_describe_method}},
  \code{\link{db_describe_organization}},
  \code{\link{db_describe_person}},
  \code{\link{db_describe_site}},
  \code{\link{db_describe_variable}}
}
\concept{describe functions}
