% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_method.R
\name{db_describe_method}
\alias{db_describe_method}
\title{Describe a new method}
\usage{
db_describe_method(db, methodname, methodcode, methodtypecv, methoddescription)
}
\arguments{
\item{db}{database connection object}

\item{methodname}{full name of method}

\item{methodcode}{short codename for method}

\item{methodtypecv}{method type from controlled vocabulary}

\item{methoddescription}{longer description}
}
\value{
TRUE if successful
}
\description{
Describe a new method
}
\examples{
\dontrun{
db_describe_method("new method", "new", "Field Activity", "doing stuf")
}
}
