% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_person.R
\name{db_describe_person}
\alias{db_describe_person}
\title{Describe a new person in the people and affiliations tables}
\usage{
db_describe_person(db = db, PersonFirstName, PersonLastName,
  AffiliationStartDate, PrimaryEmail)
}
\arguments{
\item{db}{Database connection object (must be postgresql or sqlite)}

\item{PersonFirstName}{First name of the person}

\item{PersonLastName}{Last name of the person}

\item{AffiliationStartDate}{The date (YYYY-MM-DD) on which the person became affiliated with the organization}

\item{PrimaryEmail}{The primary email address of the person}
}
\description{
Describe a new person in the people and affiliations tables
}
