% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_site.R
\name{db_describe_site}
\alias{db_describe_site}
\title{Describe a new site}
\usage{
db_describe_site(db, site_code, sitetypecv = "Site", site_name = NULL,
  site_description = NULL)
}
\arguments{
\item{db}{database connection object}

\item{site_code}{unique short code name (required)}

\item{sitetypecv}{site type from [controlled vocab](http://vocabulary.odm2.org/samplingfeaturetype/) such as "Site" or "Weather station"}

\item{site_name}{optional longer site name}

\item{site_description}{optional longer site description}
}
\value{
message that site was added
}
\description{
Describe a new site
}
\examples{
# just add site name
db <- rodm2::connect_sqlite()
db_describe_site(db, site_code = "new_site")
}
