% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_site.R
\name{db_describe_site_geom}
\alias{db_describe_site_geom}
\title{Describe a new site with geometry}
\usage{
db_describe_site_geom(db, site_code, site_geometry, sitetype, ...)
}
\arguments{
\item{db}{database connection}

\item{site_code}{unique short code name for site or maybe column of sf with site code}

\item{site_geometry}{sf object with point coordinates of site}

\item{sitetype}{term from sitetype controlled vocabulary}

\item{...}{additional parameters to db_describe_site (site_name, site_description)}
}
\value{
message that site has been added
}
\description{
Describe a new site with geometry
}
\examples{
# add a site with coordinate information
db <- rodm2::create_sqlite(connect = TRUE)
pt1 <- sf::st_sfc(sf::st_point(c(-76.503394, 38.976546)),crs = sf::st_crs(4326))
site_sf <- sf::st_sf(data.frame(site_code = "my_site", geom = pt1))
db_describe_site_geom(db, site_code = "my_code", site_geometry = site_sf, sitetype = 'Unknown')
}
