% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_variable.R
\name{db_describe_variable}
\alias{db_describe_variable}
\title{Describe a new variable}
\usage{
db_describe_variable(db, variablecode, variabletypecv, variablenamecv,
  variabledefinition = NULL, nodatavalue = -9999)
}
\arguments{
\item{db}{database connection object}

\item{variablecode}{short codename for variable}

\item{variabletypecv}{variable type from \href{http://vocabulary.odm2.org/variabletype}{controlled vocabulary}}

\item{variablenamecv}{variable name from \href{http://vocabulary.odm2.org/variablename}{controlled vocabulary}}

\item{variabledefinition}{optional longer definition}

\item{nodatavalue}{no data value}
}
\value{
TRUE if successful
}
\description{
Describe a new variable
}
\examples{
\dontrun{
db_describe_variable()
}
}
\seealso{
Other describe functions: \code{\link{db_describe_annotation}},
  \code{\link{db_describe_equipment}},
  \code{\link{db_describe_method}},
  \code{\link{db_describe_organization}},
  \code{\link{db_describe_person}},
  \code{\link{db_describe_site}}
}
\concept{describe functions}
