% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_annotation.R
\name{db_get_annotations}
\alias{db_get_annotations}
\title{Get list of annotations currently in database}
\usage{
db_get_annotations(db)
}
\arguments{
\item{db}{database connection object}
}
\value{
a dataframe with the current values in the annotation code and annotation text columns
}
\description{
Get list of annotations currently in database
}
\examples{
db <- create_sqlite(connect = TRUE)

db_describe_annotation(db,
annotationtext = "Riparian wells",
annotationtypecv = "Site group")

db_get_annotations(db)
}
