% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_equipment.R
\name{db_get_equipment}
\alias{db_get_equipment}
\title{Get list of equipment currently in database}
\usage{
db_get_equipment(db)
}
\arguments{
\item{db}{}
}
\value{
the current values in the equipmentcode column of the equipment table
}
\description{
Get list of equipment currently in database
}
\examples{
#db_get_methods(db)
}
