% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_person.R
\name{db_get_people}
\alias{db_get_people}
\title{Get list of people currently in database}
\usage{
db_get_people(db)
}
\arguments{
\item{db}{database connection object}
}
\value{
the current first and last names in the people table
}
\description{
Get list of people currently in database
}
\examples{
#db_get_methods(db)
}
