% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get_results.R
\name{db_get_results}
\alias{db_get_results}
\title{Title}
\usage{
db_get_results(db, site_code = NULL, variable_code = NULL,
  result_type = c("sample", "ts", "measurement"))
}
\arguments{
\item{result_type}{}
}
\description{
Title
}
