% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_sample.R
\name{db_get_samples}
\alias{db_get_samples}
\title{Get list of sample codes currently in database}
\usage{
db_get_samples(db)
}
\arguments{
\item{db}{database connection object}
}
\value{
the current values in the samplingfeaturecode column
}
\description{
Get list of sample codes currently in database
}
\examples{
db <- create_sqlite(connect = TRUE)
db_describe_site(db, site_code = "site1")
db_describe_sample(db, sample_code = "samp01", site_code = "site1")
db_get_samples(db)

}
