% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_variable.R
\name{db_get_variables}
\alias{db_get_variables}
\title{Get list of variables currently in database}
\usage{
db_get_variables(db)
}
\arguments{
\item{db}{}
}
\value{
the current values in the variablenamecv column
}
\description{
Get list of variables currently in database
}
\examples{
#db_get_variables(db)
}
