% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-measurement-results.R
\name{db_insert_results_m}
\alias{db_insert_results_m}
\title{Insert time series data values to ODM2 database}
\usage{
db_insert_results_m(db, datavalues, method, site_code_col = "site",
  variables, sampledmedium, processinglevel = "Raw data",
  aggregationstatistic = "Unknown", time_aggregation_interval = list(1,
  "Minute"), methodtypecv = "Instrument deployment", actionby = NULL,
  equipment_name = NULL, zlocation = NULL, zlocationunits = NULL,
  ...)
}
\arguments{
\item{db}{database connection}

\item{datavalues}{data frame with columns "Timestamp" with
POSIXct YYYY-MM-DD H:M:S format, and column names corresponding to variable names}

\item{method}{code for method used to collect data}

\item{site_code_col}{name of column with site codes. defaults to "site"}

\item{variables}{a named list of lists defining variable names, units, and columns in datavalues data frame
with format list("variablen1ame" = list(units = 'unitsname', column = 'colname', dataqualitycol = 'qualcode'))}

\item{sampledmedium}{term from controlled vocabulary for medium sampled eg. Air, Water, Soil}

\item{processinglevel}{code for processing level. will be added to processinglevels table if new.
defaults to "Raw data'.}

\item{methodtypecv}{}

\item{actionby}{(optional) the person who performed the action}

\item{equipment_name}{(optional) the equipment used to collect the data}

\item{zlocation}{(optional) z location offset}

\item{zlocationunits}{(optional but required if zlocation is set) name of units of z location offset}

\item{...}{parameters to pass to various db_describe_ functions}
}
\value{
true if successful
}
\description{
Insert time series data values to ODM2 database
}
\examples{
\dontrun{
# make database and data frame
db <- rodm2::create_sqlite(connect = TRUE)
mrv <- data.frame(Timestamp = "2018-06-27 13:55:00",
"vwc" = 68.3, site = "BB2")

db_insert_results_m(db = db, datavalues = mrv,
    method = "soilmoisture", site_code_col = "site",
    variables = list("Volumetric water content" = list("vwc"," Percent")),
    sampledmedium = "Soil")
}

}
