% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-sample-profile-results.R
\name{db_insert_results_samples_profile}
\alias{db_insert_results_samples_profile}
\title{Insert sample result data values to ODM2 database}
\usage{
db_insert_results_samples_profile(db, datavalues, field_method = NULL,
  lab_method = NULL, lab_method_col = NULL, variables, sampledmedium,
  site_code_col = "Site", sample_code_col = "Sample",
  processinglevel = "Raw data", aggregationstatistic = "Unknown",
  time_aggregation_interval = list(1, "Minute"), field_actionby = NULL,
  field_equipment_name = NULL, lab_actionby = NULL,
  lab_equipment_name = NULL, zlocation_col = NULL,
  zinterval_col = NULL, zlocationunits, ...)
}
\arguments{
\item{db}{database connection}

\item{datavalues}{data frame with data to upload columns "Timestamp" with
YYYY-MM-DD H:M:S format, and column names corresponding to variable names}

\item{field_method}{short description of field method to collect new samples. Dont include for new data about existing samples.}

\item{lab_method}{short description of lab method. only use if lab_method_col is NULL}

\item{lab_method_col}{specify the column name of the lab method. only use if lab method is NULL}

\item{variables}{a named list of lists defining variable names, units, and columns in datavalues data frame
with format list("variablen1ame" = list(units = 'unitsname', column = 'colname', dataqualitycol = 'qualcode'))}

\item{sampledmedium}{term from controlled vocabulary for medium sampled eg. Air, Water, Soil}

\item{site_code_col}{name of column with site codes. defaults to "Site"}

\item{sample_code_col}{name of column in input data frame with sample ID. defeaults to "Sample"}

\item{processinglevel}{code for processing level. will be added to processinglevels table if new.
defaults to "Raw data'.}

\item{aggregationstatistic}{defaults to unknown}

\item{time_aggregation_interval}{defaults to unknown}

\item{field_actionby}{optional first name of lead for field action, must already be in people table}

\item{field_equipment_name}{optional code name of equipment used for field action, must already be in equipment table}

\item{lab_actionby}{optional first name of lead for lab action, must already be in people table}

\item{lab_equipment_name}{optional code name of equipment used for lab action, must already be in equipment table}

\item{zlocation_col}{specify column name of the zlocation}

\item{zinterval_col}{specify column name of the z aggregation interval}

\item{zlocationunits}{name of units of z location offset FROM CONTROLLED VOCAB}

\item{...}{parameters to pass to various db_describe_ functions}
}
\value{
true if successful
}
\description{
Insert sample result data values to ODM2 database
}
\examples{
\dontrun{
# make database and data frame
db <- rodm2::create_sqlite(connect = TRUE)
mrv <- data.frame(Timestamp = "2018-06-27 13:55:00",
"pH" = 4.5, Site = "BB2", Sample = "CEM 1", stringsAsFactors = FALSE)

db_insert_results_sample(db = db, datavalues = soil_ph,
    field_method = "1 inch 30 cm soil core",
    lab_method = "Soil pH CEM",
    variables = list("pH" = list(column = "pH", units = "pH")),
    sampledmedium = "Soil")
}

}
\seealso{
Other insert data functions: \code{\link{db_insert_results_m}},
  \code{\link{db_insert_results_samples}},
  \code{\link{db_insert_results_ts}}
}
\concept{insert data functions}
