% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_insert_sample_relations.R
\name{db_insert_sample_relations}
\alias{db_insert_sample_relations}
\title{Insert new sample relationships to the related features table}
\usage{
db_insert_sample_relations(new_related_samples, lhc, relationshiptypecv,
  rhc)
}
\arguments{
\item{new_related_samples}{data frame with one row for each new relationship}

\item{lhc}{column with sample codes}

\item{relationshiptypecv}{relationship type from controlled vocabulary}

\item{rhc}{column with sample codes of related features}
}
\value{
true if successful
}
\description{
Insert new sample relationships to the related features table
}
\examples{
\dontrun{
db_insert_sample_relations(new_related_samples, "hs_sample", "source_air", "Is part of")
}
}
