% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_insert_samples.R
\name{db_insert_samples}
\alias{db_insert_samples}
\title{Insert new samples from a new_samples dataframe}
\usage{
db_insert_samples(x)
}
\arguments{
\item{x}{}
}
\value{
TRUE if successful
}
\description{
Insert new samples from a new_samples dataframe
}
\examples{
\dontrun{
db_insert_samples(1)
}
}
