% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_insert_sites.R
\name{db_insert_sites}
\alias{db_insert_sites}
\title{Insert new sites from a table}
\usage{
db_insert_sites(db, new_sites, SamplingFeatureTypeCV = "site",
  code_name_column = "SamplingFeatureCode", long_name_column = NULL,
  description_column = NULL, geometry_wkt_column = NULL,
  elevation_column = NULL, add_uuids = TRUE)
}
\arguments{
\item{x}{}
}
\value{
TRUE if successful
}
\description{
Insert new sites from a table
}
\examples{
\dontrun{
db_insert_sites(new_sites)
}
}
