% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_insert_sites_latlon.R
\name{db_insert_sites_latlon}
\alias{db_insert_sites_latlon}
\title{Insert new sites from a new_sites table}
\usage{
db_insert_sites_latlon(x)
}
\arguments{
\item{x}{}
}
\value{
TRUE if successful
}
\description{
Insert new sites from a new_sites table
}
\examples{
\dontrun{
db_insert_sites_latlon(new_sites)
}
}
