% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sqlite.R
\name{disconnect_sqlite}
\alias{disconnect_sqlite}
\title{Disconnect from connect ODM2 sqlite database}
\usage{
disconnect_sqlite(db = db)
}
\arguments{
\item{db}{database connection object}
}
\value{
A message about how to reconnect
}
\description{
Disconnect from connect ODM2 sqlite database
}
\details{
This function is a simple wrapper around dbDisconnect
}
\examples{
db <- create_sqlite(dir = tempdir(), connect = TRUE)
disconnect_sqlite(db)

}
