% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cv_names.R
\name{get_cv_names}
\alias{get_cv_names}
\title{Print a list of controlled vocabulary terms}
\usage{
get_cv_names(cvtype)
}
\arguments{
\item{cvtype}{the name of the controlled vocab eg. "resulttype" or "methodtype"}
}
\value{
a vector of the Names column from the CV_%cvtype% table
}
\description{
Print a list of controlled vocabulary terms
}
\examples{
get_cv_names("resulttype")
resultterms <- get_cv_names("resulttype")
}
