% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_codes.R
\name{get_site_codes}
\alias{get_site_codes}
\title{Shiny gadget to get site codes of interest}
\usage{
get_site_codes(db)
}
\arguments{
\item{db}{database connection object}
}
\value{
selected site codes
}
\description{
Shiny gadget to get site codes of interest
}
\details{
Use site group annotations, site types, and
site relations to identify site codes to
use in db_get_results()
}
\examples{
# my_sites <- get_site_codes(db)

}
\seealso{
Other interactive helpers: \code{\link{get_variable_codes}},
  \code{\link{make_vars_list}}
}
\concept{interactive helpers}
