% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable_codes.R
\name{get_variable_codes}
\alias{get_variable_codes}
\title{Shiny gadget to get variable codes of interest}
\usage{
get_variable_codes(db)
}
\arguments{
\item{db}{database connection object}
}
\value{
selected variable codes
}
\description{
Shiny gadget to get variable codes of interest
}
\examples{
# my_vars <- get_variable_codes(db)

}
\seealso{
Other interactive helpers: \code{\link{make_vars_list}}
}
\concept{interactive helpers}
