% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_new_sf_annotation.R
\name{insert_new_annotation}
\alias{insert_new_annotation}
\title{Insert new annotation into annotations table}
\usage{
insert_new_annotation(annotationtypecv, annotationtext)
}
\arguments{
\item{annotationtext}{}
}
\description{
Insert new annotation into annotations table
}
\examples{

}
