% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_new_sf_annotation.R
\name{insert_new_sf_annotation}
\alias{insert_new_sf_annotation}
\title{Insert new samplingfeature annotation into samplingfeatureannotations table}
\usage{
insert_new_sf_annotation(samplingfeaturecode, annotationtext)
}
\arguments{
\item{annotationtext}{}
}
\description{
Insert new samplingfeature annotation into samplingfeatureannotations table
}
