% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{insert_processinglevel}
\alias{insert_processinglevel}
\title{Insert new processinglevel}
\usage{
insert_processinglevel(db, processinglevel)
}
\arguments{
\item{db}{database connection}

\item{processinglevel}{processinglevel code}
}
\value{
message if succesful
}
\description{
Insert new processinglevel
}
\examples{
#insert_processinglevel(db, "Quality controlled data")

}
