% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_gadget.R
\name{make_vars_list}
\alias{make_vars_list}
\title{Make variables list}
\usage{
make_vars_list(data)
}
\arguments{
\item{data}{data frame wtih Timestamp column and other data}
}
\value{
nested list for each column in data describing variable name and units
}
\description{
Make variables list
}
\examples{
\dontrun{
vars_list <- make_vars_list(ts_data)
}

}
