% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%>\%}
\alias{\%>\%}
\alias{get_site_names_like}
\title{Search for site names with a pattern}
\usage{
get_site_names_like(x, db, ...)
}
\arguments{
\item{x}{pattern to search for}

\item{db}{database connection}

\item{...}{extra parameters to pass to agrep}
}
\value{
matching elements if the samplingfeaturecode vector
}
\description{
Uses agrep to search the samplingfeaturecode column of the samplingfeature table
}
\examples{
\dontrun{
get_site_names_like("weather")
# Pass extra arguments to agrep
get_site_names_like("weather", max.distance = 2)
}
}
