% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodm2-package.R
\docType{package}
\name{rodm2-package}
\alias{rodm2}
\alias{rodm2-package}
\title{rodm2: Use an ODM2 Database in R}
\description{
Organize site, sample, and sensor data with R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/khondula/rodm2}
  \item Report bugs at \url{hhttps://github.com/khondula/rodm2/issues}
}

}
\author{
\strong{Maintainer}: Kelly Hondula \email{khondula@sesync.org}

Authors:
\itemize{
  \item Nate Jones \email{njones@umd.edu}
}

}
\keyword{internal}
