% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_checkers.R
\name{check_medium_cv}
\alias{check_medium_cv}
\title{Check that sampled medium term is in cv}
\usage{
check_medium_cv(sampledmedium)
}
\arguments{
\item{sampledmedium}{medium to check}
}
\value{
sampled medium
}
\description{
Check that sampled medium term is in cv
}
\examples{
# medium <- check_medium_cv()
}
