% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_checkers.R
\name{check_methodtype_cv}
\alias{check_methodtype_cv}
\title{Check that method type is in cv}
\usage{
check_methodtype_cv(methodtypecv)
}
\arguments{
\item{methodtypecv}{method type to check}
}
\value{
method type
}
\description{
Check that method type is in cv
}
\examples{
# methodcode <- check_methodtype_cv()
}
