% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sqlite.R
\name{create_sqlite}
\alias{create_sqlite}
\title{Make a new ODM2 sqlite database}
\usage{
create_sqlite(dir = ".", filename = "odm2", connect = FALSE)
}
\arguments{
\item{dir}{Name of existing directory for the sqlite database file to be created in. Defaults to current directory}

\item{filename}{Filename for sqlite database file. Defaults to odm2, creating odm2.sqlite}

\item{connect}{whether or not to return the database connection object}
}
\description{
Make a new ODM2 sqlite database
}
\details{
This function copies a template sqlite file with the ODM2 schema and controlled vocabularies to the specified location.
}
\examples{
create_sqlite(dir = tempdir())
dblite <- DBI::dbConnect(RSQLite::SQLite(), "odm2.sqlite")
# OR
dblite <- create_sqlite(connect = TRUE)
}
