% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sqlite_db.R
\name{create_sqlite}
\alias{create_sqlite}
\title{Put template sqlite database within a data folder}
\usage{
create_sqlite(dir = "data")
}
\arguments{
\item{dir}{Directory for the sqlite database file to be created in. Defaults to `data`}
}
\description{
Put template sqlite database within a data folder
}
\examples{
\dontrun{
create_sqlite()
}
}
