% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sqlite.R
\name{create_sqlite}
\alias{create_sqlite}
\title{Make a new ODM2 sqlite database}
\usage{
create_sqlite(dir = ".", filename = "odm2.sqlite", connect = FALSE)
}
\arguments{
\item{dir}{Name of existing directory for the sqlite database file to be created in. Defaults to current directory}

\item{filename}{Filename for sqlite database file. Defaults to "odm2", making a file called odm2.sqlite}

\item{connect}{Whether or not to return the database connection object}
}
\value{
Returns database connecton object if connect = TRUE, otherwise just creates file.
}
\description{
Make a new ODM2 sqlite database
}
\details{
This function copies a template sqlite file with the ODM2 schema, controlled vocabularies,
and units to the specified location.
}
\examples{
db <- create_sqlite(dir = tempdir(), connect = TRUE)

}
