% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sqlite_db.R
\name{create_sqlite}
\alias{create_sqlite}
\title{Make a new ODM2 sqlite database}
\usage{
create_sqlite(dir = "data", filename = "odm2")
}
\arguments{
\item{dir}{Directory for the sqlite database file to be created in. Defaults to `data`}

\item{filename}{Filename for sqlite database file. Defaults to `odm2`}
}
\description{
Make a new ODM2 sqlite database
}
\details{
This function copies a template sqlite file with the ODM2 schema and controlled vocabularies to the specified location.
}
\examples{
\dontrun{
create_sqlite()
}
}
