% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_add_relation.R
\name{db_add_relation}
\alias{db_add_relation}
\title{Add relationships to related features table}
\usage{
db_add_relation(db, edgelist, from_column = "SamplingFeatureCode",
  to_column = "RelatedFeatureCode", relationtype_column = NULL,
  relationtype = "isChildOf")
}
\arguments{
\item{db}{database connection object}

\item{edgelist}{dataframe with new relationships to add. see details}

\item{from_column}{name of first column with feature codes to relate}

\item{to_column}{name of second column with feature codes to relate}

\item{relationtype_column}{name of column with relationship type from [controlled vocabulary](http://vocabulary.odm2.org/relationshiptype/)}

\item{relationtype}{relationship type to use if no relationship type column. defaults to "is child of"}
}
\value{
message with how many rows added
}
\description{
Add relationships to related features table
}
\details{
Edgelist should be a data frame with 3 columns and one row for each new
relationship to establish with the features to connect and the type of relationship.
Specify the names of columns corresponding to the features using the 'from_column' and
'to_column' arguments, and the name of the column with relationship types with
'relationtype_column'. A two column data frame can be supplied along with a relationtype
to use for all of the new relationships.
}
\examples{
# df <- data.frame(
# "SamplingFeatureCode" = "subsite1",
# "RelatedFeatureCode" = "site1")
# db_add_relation(db, edgelist = df)
}
