% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_add_variable.R
\name{db_add_variable}
\alias{db_add_variable}
\title{Insert a new variable}
\usage{
db_add_variable(variabletypecv, variablecode, variablenamecv,
  variabledefinition, nodatavalue = -9999)
}
\arguments{
\item{variabletypecv}{}

\item{variablecode}{}

\item{variablenamecv}{}

\item{variabledefinition}{}

\item{nodatavalue}{}
}
\description{
Insert a new variable
}
\examples{
\dontrun{
db_add_variable()
}
}
