% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_annotation.R
\name{db_describe_annotation}
\alias{db_describe_annotation}
\title{Add a new annotation text}
\usage{
db_describe_annotation(db, annotationtext = NULL,
  annotationtypecv = NULL, annotationcode = NULL)
}
\arguments{
\item{db}{database connection object}

\item{annotationtext}{text string of annotation.}

\item{annotationtypecv}{annotation type from \href{http://vocabulary.odm2.org/annotationtype/}{controlled vocab}}

\item{annotationcode}{Optional short codename for annotation}
}
\value{
message if successful
}
\description{
Add a new annotation text
}
\details{
Use this function to add new annotations such as the name of a Site group.
Annotation text is required but if an annotation code is provided then that
will be used as the annotation text as well.
}
\examples{
db <- create_sqlite(connect = TRUE)

db_describe_annotation(db,
annotationtext = "Riparian wells",
annotationtypecv = "Site group")

db_describe_annotation(db,
annotationtext = "January sampling campaign",
annotationtypecv = "Specimen group",
annotationcode = "Jan")

}
\seealso{
Other describe functions: \code{\link{db_describe_equipment}},
  \code{\link{db_describe_method}},
  \code{\link{db_describe_organization}},
  \code{\link{db_describe_person}},
  \code{\link{db_describe_site}},
  \code{\link{db_describe_variable}}
}
\concept{describe functions}
