% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_annotation.R
\name{db_describe_annotation}
\alias{db_describe_annotation}
\title{Add a new annotation text}
\usage{
db_describe_annotation(db, annotationtypecv, annotationtext)
}
\arguments{
\item{db}{database connection object}

\item{annotationtypecv}{annotation type from controlled vocab}

\item{annotationtext}{text string of annotation}
}
\value{
TRUE if successful
}
\description{
Add a new annotation text
}
\examples{
\dontrun{
db_describe_annotaion("samplingfeature annotation", "primary sites")
}

}
