% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_method.R
\name{db_describe_method}
\alias{db_describe_method}
\title{Describe a new method}
\usage{
db_describe_method(db, methodname, methodcode, methodtypecv,
  methoddescription = NULL)
}
\arguments{
\item{db}{database connection object}

\item{methodname}{full name of method}

\item{methodcode}{short codename for method}

\item{methodtypecv}{method type from controlled vocabulary}

\item{methoddescription}{optional longer description}
}
\value{
TRUE if successful
}
\description{
Describe a new method
}
\examples{
db <- create_sqlite(connect = TRUE)

db_describe_method(db,
methodname = "new method",
methodcode = "method1",
methodtypecv =  "fieldActivity")
}
\seealso{
Other describe functions: \code{\link{db_describe_annotation}},
  \code{\link{db_describe_equipment}},
  \code{\link{db_describe_organization}},
  \code{\link{db_describe_person}},
  \code{\link{db_describe_site}},
  \code{\link{db_describe_variable}}
}
\concept{describe functions}
