% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_organization.R
\name{db_describe_organization}
\alias{db_describe_organization}
\title{Add a new entry to the Organizations table}
\usage{
db_describe_organization(db = db, OrganizationTypeCV, OrganizationCode,
  OrganizationName)
}
\arguments{
\item{db}{database connection object to sqlite or postgresql database with odm2 schema}

\item{OrganizationTypeCV}{Controlled vocab term defining the type of organization (e.g., government agency, university, etc.)}

\item{OrganizationCode}{A text code identifying the Organization (e.g., USGS)}

\item{OrganizationName}{The full text name of the organization}
}
\value{
message that your entry has been added
}
\description{
Add a new entry to the Organizations table
}
\examples{
# connect_sqlite(dir = ".")
# db <- DBI::dbConnect(RSQLite::SQLite(), "odm2.sqlite")
# db_describe_organization(db, "Manufacturer", "Onset", "Onset")
}
\seealso{
Other describe functions: \code{\link{db_describe_annotation}},
  \code{\link{db_describe_equipment}},
  \code{\link{db_describe_method}},
  \code{\link{db_describe_person}},
  \code{\link{db_describe_site}},
  \code{\link{db_describe_variable}}
}
\concept{describe functions}
