% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_person.R
\name{db_describe_person}
\alias{db_describe_person}
\title{Describe a new person in the people and affiliations tables}
\usage{
db_describe_person(db = db, PersonFirstName, PersonLastName,
  AffiliationStartDate = Sys.Date(), PrimaryEmail)
}
\arguments{
\item{db}{Database connection object (must be postgresql or sqlite)}

\item{PersonFirstName}{First name of the person}

\item{PersonLastName}{Last name of the person}

\item{AffiliationStartDate}{The date (YYYY-MM-DD) on which the person became affiliated with the organization. Default is todays date}

\item{PrimaryEmail}{The primary email address of the person}
}
\value{
message that person was entered into database
}
\description{
Describe a new person in the people and affiliations tables
}
\examples{
# db <- connect_sqlite(dir = ".")
#db_describe_person(db = db, PersonFirstName = "Wendy",
 #   PersonLastName = "Wetland",
  #  AffiliationStartDate = "2018-01-01",
   # PrimaryEmail = "wendy 'at' swamps.edu")
}
\seealso{
Other describe functions: \code{\link{db_describe_annotation}},
  \code{\link{db_describe_equipment}},
  \code{\link{db_describe_method}},
  \code{\link{db_describe_organization}},
  \code{\link{db_describe_site}},
  \code{\link{db_describe_variable}}
}
\concept{describe functions}
