% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_site.R
\name{db_describe_site}
\alias{db_describe_site}
\title{Describe a new site}
\usage{
db_describe_site(db, site_code, sitetypecv = "Site", site_name = NULL,
  site_description = NULL)
}
\arguments{
\item{db}{database connection object}

\item{site_code}{unique short code name (required)}

\item{sitetypecv}{samplingfeature type from
\href{http://vocabulary.odm2.org/samplingfeaturetype/}{controlled vocab}, defaults to "Site"}

\item{site_name}{optional longer site name}

\item{site_description}{optional longer site description}
}
\value{
message that site was added
}
\description{
Describe a new site
}
\examples{
db <- create_sqlite(connect = TRUE)
db_describe_site(db, site_code = "site1")
}
\seealso{
Other describe functions: \code{\link{db_describe_annotation}},
  \code{\link{db_describe_equipment}},
  \code{\link{db_describe_method}},
  \code{\link{db_describe_organization}},
  \code{\link{db_describe_person}},
  \code{\link{db_describe_variable}}
}
\concept{describe functions}
