% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_variable.R
\name{db_describe_variable}
\alias{db_describe_variable}
\title{Describe a new variable}
\usage{
db_describe_variable(db, variabletypecv, variablecode, variablenamecv,
  variabledefinition, nodatavalue = -9999)
}
\arguments{
\item{db}{database connection object}

\item{variabletypecv}{variable type from controlled vocab}

\item{variablecode}{short codename for variable}

\item{variablenamecv}{variable name from controlled vocab}

\item{variabledefinition}{longer definition}

\item{nodatavalue}{no data value}
}
\value{
TRUE if successful
}
\description{
Describe a new variable
}
\examples{
\dontrun{
db_describe_variable()
}
}
