% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_method.R
\name{db_get_methods}
\alias{db_get_methods}
\title{Get list of methods currently in database}
\usage{
db_get_methods(db)
}
\arguments{
\item{db}{}
}
\value{
the current values in the methodname column of the methods table
}
\description{
Get list of methods currently in database
}
\examples{
#db_get_methods(db)
}
