% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_method.R, R/db_describe_variable.R
\name{db_get_methods}
\alias{db_get_methods}
\alias{db_get_methods}
\title{Get list of methods currently in database}
\usage{
db_get_methods(db)

db_get_methods(db)
}
\arguments{
\item{db}{}

\item{db}{}
}
\value{
the current values in the methodname column of the methods table

the current values in the variablenamecv column
}
\description{
Get list of methods currently in database

Get list of variables currently in database
}
\examples{
#db_get_methods(db)
#db_get_variables(db)
}
