% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get_results.R
\name{db_get_results}
\alias{db_get_results}
\title{Query result data}
\usage{
db_get_results(db, site_code = NULL, variable_code = NULL,
  result_type = c("sample", "ts", "measurement"))
}
\arguments{
\item{db}{database connection object}

\item{site_code}{a string vector with codes of sites to get data for. default NULL returns all sites}

\item{variable_code}{a string vector with codes of variables to get data for. default NULL returns all variables}

\item{result_type}{a string vector of result types to include: "ts" for time series results, "sample" for sample-based results, "measurement" for measurements}
}
\value{
a named list of dataframes corresponding to each result type.
}
\description{
Retrieve data associated with one or more sites, variables, and result types
}
\examples{
# db_get_results(db, result_type = "ts")
}
