% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_site.R
\name{db_get_samples}
\alias{db_get_samples}
\title{Get list of sample codes currently in database}
\usage{
db_get_samples(db)
}
\arguments{
\item{db}{}
}
\value{
the current values in the samplingfeaturecode column
}
\description{
Get list of sample codes currently in database
}
\examples{
#db_get_samples(db)
}
