% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_describe_site.R
\name{db_get_sites}
\alias{db_get_sites}
\title{Get list of sampling feature codes currently in database}
\usage{
db_get_sites(db)
}
\arguments{
\item{db}{database connection object}
}
\value{
Sampling feature codes for everything that is not sampling feature type 'Specimen'
}
\description{
Get list of sampling feature codes currently in database
}
\examples{
db <- create_sqlite(connect = TRUE)
db_describe_site(db, site_code = "site1")
db_get_sites(db)

}
