% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get_water_level_ts.R
\name{db_get_water_level_ts}
\alias{db_get_water_level_ts}
\title{Function to retrieve time series data with variablecode waterLevel from a specified sampling feature}
\usage{
db_get_water_level_ts(db, site_code)
}
\arguments{
\item{db}{database connection object}

\item{site_code}{site code to get water level time series for}
}
\value{
time series
}
\description{
Function to retrieve time series data with variablecode waterLevel from a specified sampling feature
}
