% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_insert_measurements.R
\name{db_insert_measurements_sm}
\alias{db_insert_measurements_sm}
\title{Insert new soil moisture measurements}
\usage{
db_insert_measurements_sm(x)
}
\arguments{
\item{x}{}
}
\value{
TRUE if successful
}
\description{
Insert new soil moisture measurements
}
\examples{
\dontrun{
db_insert_measurements_sm(1)
# insert all data from a dataframe called new_sm
sapply(1:nrow(new_sm), db_insert_measurements_sm)
}
}
