% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cv_terms.R
\name{get_cv_terms}
\alias{get_cv_terms}
\title{Print a list of controlled vocabulary terms}
\usage{
get_cv_terms(cvtype, quietly = FALSE)
}
\arguments{
\item{cvtype}{the name of the controlled vocab eg. "resulttype" or "methodtype"}

\item{quietly}{whether to print in console or not}
}
\value{
a vector of the Names column from the controlled vocab table
}
\description{
Print a list of controlled vocabulary terms
}
\examples{
get_cv_terms("resulttype")
resultterms <- get_cv_terms("resulttype")
}
