% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vars_list.R
\name{make_vars_list}
\alias{make_vars_list}
\title{Make variables list}
\usage{
make_vars_list(data, data_colnames = setdiff(names(data), c("Timestamp",
  "Sample", "Site")))
}
\arguments{
\item{data}{data frame with columns of data to describe}

\item{data_colnames}{a string vector of column names with data values.
by default all columns except "Timestamp", "Sample", and "Site" if they are included.}
}
\value{
nested list for each column in data describing variable name and units
}
\description{
Make variables list
}
\details{
This function creates the object needed for the variables argument
in db_insert_results_m, db_insert_results_ts, db_insert_results_samples,
db_insert_results_samples_profile, based on a data frame with the data to insert.
The data frame supplied should include only the columns with data and optionally a
column called Timestamp. For each column of data, provide a unique codename for the
variable as well as the \href{http://vocabulary.odm2.org/variablename}{term}
and \href{http://vocabulary.odm2.org/units}{units} from the controlled vocabularies. By
default the column name will be used as the variable code.
}
\examples{
\dontrun{
ts <- data.frame(
 Timestamp = c(
 "2018-06-27 13:45:00",
 "2018-06-27 13:55:00"),
  "wl" = c(1, 1.5))
vars_list <- make_vars_list(ts)
}

}
\seealso{
Other interactive helpers: \code{\link{get_site_codes}},
  \code{\link{get_variable_codes}}
}
\concept{interactive helpers}
