% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%>\%}
\alias{\%>\%}
\alias{get_site_names_like}
\alias{\%>\%}
\alias{check_samplingfeaturecodes}
\title{Search for site names with a pattern}
\usage{
get_site_names_like(x, ...)

check_samplingfeaturecodes(new_codes)
}
\arguments{
\item{...}{extra parameters to pass to agrep}

\item{new_codes}{vector of names to compare with samplingfeaturecode vector}

\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}
}
\value{
matching elements if the samplingfeaturecode vector

data frame with new_codes and in_db column
}
\description{
Uses agrep to search the samplingfeaturecode column of the samplingfeature table

Returns a data frame specifying whether each element of new_codes exists in samplingfeatures
}
\examples{
\dontrun{
get_site_names_like("weather")
# Pass extra arguments to agrep
get_site_names_like("weather", max.distance = 2)
}
\dontrun{
#
check_samplingfeaturecodes(new_codes = new_sites$sites)
}

}
